+++ gargantua/utils/tornado.py
@@ -117,7 +117,6 @@ class AuthHandlerMixin():
 
     def get_current_user(self):
         logger.debug('get_current_user')
-        return True
 
         try:
             cli_token = self.get_cookie('token')
@@ -190,7 +189,10 @@ class HttpErrorMixin():
     """
 
     def parse_err(self, err):
-        return ' '.join(str(_) for _ in err.args)
+        if isinstance(err, BaseException):
+            return ' '.join(str(_) for _ in err.args)
+        else:
+            return str(err)
 
     def http_400_bad_request(self, **kwargs):
         if 'err' in kwargs: